+++ src/Sulu/Bundle/ContentBundle/Tests/Functional/Controller/NodeControllerTest.php
@@ -207,7 +207,7 @@ class NodeControllerTest extends SuluTestCase
         $this->assertEquals(200, $client->getResponse()->getStatusCode());
 
         $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data);
-        $this->assertequals(409, $client->getResponse()->getStatusCode());
+        $this->assertEquals(409, $client->getResponse()->getStatusCode());
     }
 
     public function testGet()
@@ -811,6 +811,92 @@ class NodeControllerTest extends SuluTestCase
         $client->request('GET', '/api/nodes/' . $response['id'] . '?language=en', $data);
         $response = json_decode($client->getResponse()->getContent(), true);
         $this->assertEquals(1, $response['nodeState']);
+    }
+
+    public function testPutWithValidHash()
+    {
+        $data = [
+            'title' => 'Testtitle',
+            'template' => 'default',
+            'url' => '/test',
+        ];
+
+        $client = $this->createAuthenticatedClient();
+
+        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data);
+        $response = json_decode($client->getResponse()->getContent(), true);
+
+        $client->request('GET', '/api/nodes/' . $response['id'] . '?language=en', $data);
+        $response = json_decode($client->getResponse()->getContent(), true);
+
+        $client->request(
+            'PUT',
+            '/api/nodes/' . $response['id'] . '?webspace=sulu_io&language=en&state=2',
+            array_merge(['_hash' => $response['_hash']], $data)
+        );
+
+        $this->assertEquals(200, $client->getResponse()->getStatusCode());
+    }
+
+    public function testPutWithInvalidHash()
+    {
+        $data = [
+            'title' => 'Testtitle',
+            'template' => 'default',
+            'url' => '/test',
+        ];
+
+        $client = $this->createAuthenticatedClient();
+
+        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data);
+        $response = json_decode($client->getResponse()->getContent(), true);
+        $id = $response['id'];
+
+        $client->request(
+            'PUT',
+            '/api/nodes/' . $id . '?webspace=sulu_io&language=en&state=2',
+            array_merge(['_hash' => md5('wrong-hash')], $data)
+        );
+
+        $this->assertEquals(409, $client->getResponse()->getStatusCode());
+        $response = json_decode($client->getResponse()->getContent(), true);
+        $this->assertEquals(1102, $response['code']);
+
+        $client->request(
+            'PUT',
+            '/api/nodes/' . $id . '?webspace=sulu_io&language=en&state=2&force=true',
+            array_merge(['_hash' => md5('wrong-hash')], $data)
+        );
+
+        $this->assertEquals(200, $client->getResponse()->getStatusCode());
+    }
+
+    public function testPutWithAlreadyExistingUrl()
+    {
+        $client = $this->createAuthenticatedClient();
+
+        $data = [
+            'title' => 'Testtitle',
+            'template' => 'default',
+            'url' => '/test',
+        ];
+        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data);
+        $this->assertEquals(200, $client->getResponse()->getStatusCode());
+
+        $data['url'] = '/test2';
+        $client->request('POST', '/api/nodes?webspace=sulu_io&language=en', $data);
+        $this->assertEquals(200, $client->getResponse()->getStatusCode());
+        $response = json_decode($client->getResponse()->getContent(), true);
+
+        $data['url'] = '/test';
+        $client->request(
+            'PUT',
+            '/api/nodes/' . $response['id'] . '?webspace=sulu_io&language=en&state=2',
+            $data
+        );
+        $this->assertEquals(409, $client->getResponse()->getStatusCode());
+        $response = json_decode($client->getResponse()->getContent(), true);
+        $this->assertEquals(1103, $response['code']);
     }
 
     private function import($fileName)